requirements.txt
fastapi
uvicorn
transformers
torch
sentence-transformers
faiss-cpu
playwright
beautifulsoup4
numpy
